Extract `SearchField` to a separate file.
